+++ lib/augmentConfig.js
@@ -4,9 +4,8 @@ const _ = require('lodash');
 const configurationError = require('./utils/configurationError');
 const getModulePath = require('./utils/getModulePath');
 const globjoin = require('globjoin');
-const normalizeRuleSettings = require('./normalizeRuleSettings');
+const normalizeAllRuleSettings = require('./normalizeAllRuleSettings');
 const path = require('path');
-const rules = require('./rules');
 
 /** @typedef {import('stylelint').StylelintConfigPlugins} StylelintConfigPlugins */
 /** @typedef {import('stylelint').StylelintConfigProcessor} StylelintConfigProcessor */
@@ -290,37 +289,6 @@ function addPluginFunctions(config) {
 	}, /** @type {{[k: string]: Function}} */ ({}));
 
 	config.pluginFunctions = pluginFunctions;
-
-	return config;
-}
-
-/**
- * @param {StylelintConfig} config
- * @return {StylelintConfig}
- */
-function normalizeAllRuleSettings(config) {
-	/** @type {StylelintConfigRules} */
-	const normalizedRules = {};
-
-	if (!config.rules) return config;
-
-	Object.keys(config.rules).forEach((ruleName) => {
-		const rawRuleSettings = _.get(config, ['rules', ruleName]);
-
-		const rule = rules[ruleName] || _.get(config, ['pluginFunctions', ruleName]);
-
-		if (!rule) {
-			normalizedRules[ruleName] = [];
-		} else {
-			normalizedRules[ruleName] = normalizeRuleSettings(
-				rawRuleSettings,
-				ruleName,
-				_.get(rule, 'primaryOptionArray'),
-			);
-		}
-	});
-
-	config.rules = normalizedRules;
 
 	return config;
 }
